+++ test/mock-data/common-data-sets.js
@@ -1,3 +1,4 @@
 
 exports.TrueDataSet = [true, 'on', 'On', 'ON', 1, 'y', 'Y', 'yes', 'YES', 'true', 'TRUE', '1']
 exports.FalseDataSet = [false, 'off', 'Off', 'OFF', 0, 'n', 'N', 'no', 'NO', 'false', 'FALSE', '0']
+exports.NotBooleanDataSet = ['ANC', '3453', 'xxx', 567, 3.3, [], {}, { x: 1 }, () => { }, class { }]